Use presence of start_time instead of to_tag to evaluate expired session
In the dispatcher, look at this value to evaluate if accounting should be
performed for the expired session and if the dialog module of OpenSIPS
should be notified. If a re-INVITE is the first time the relay hears
about the session, it will have already have a to_tag. The star_time only
gets set if the relay saw SDP from both the caller and the callee.